Problem solved - The two cubes and the boolean modifier
The problem was that the cubes were temporarily hidden. This is why they showed up in the viewport. To have the effect permanent, we need to apply the modifier. Now, when the main cube (body) is clicked, the boolean modifier is not present. This is because I selected the apply option on the boolean modifier by clicking on the arrow beside the camera on the boolean modifier of the cube. After applying it, I deleted the two cubes and the effect became permanent.