Both tyres completed
I used the mirror modifier to add the second tyre. The mirror modifier worked for the tyre frame and the gears, however, it didn't work for the back wall kind of thing I had created. Due to this issue, I had to copy-paste it on to the other side and adjust the position